ekb-platform/backend/python_components/requirements.txt
fastapi
uvicorn
langchain
psycopg2-binary
openpipe
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
authlib
pgvector
sentence-transformers
